customer-support-agent/requirements.txt
strands-agents
strands-agents-tools
bedrock-agentcore
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0
boto3>=1.34.0
